Move tests to "quirks" folder and split tests to 3 files
Matches the directory format of the source
