+++ package.json
@@ -7,6 +7,6 @@
         "validator": "^13.0.0"
     },
     "scripts": {
-        "test": "mocha"
+        "test": "mocha --recursive"
     }
 }
